213 feature npm build fix (#222)
* moved file

* removed 'e' since it's never used

* modev config array out of useEffects and added locked to dependecy array

* updated text to swedish

* Updated winning logic text

* Updated button text

* added classname prop

* removed unused props

* added classname

* moved winner button

* removed unused imports and props

* fixed for build

* removed unused import

* removed jwt display

* removed export

* removed unused component

* removed old background image logic

* split up GamContext and hook, updated import links

* removed unecessary file

* removed unused hook

* added error display

* removed unused e

* split up game context

* updated

* added decoded token interface

* removed test files

* updated typ to unknown

* set return to null

* update vite.config.js

* removed unused component

* removed

* fixed window
