+++ src/app/components/EndScreen.tsx
@@ -2,28 +2,11 @@
 import Button from './Button';
 import styled from '@emotion/styled';
 import { useHighScore } from './useHighScore';
-import { processReward } from "../services/transactionService";
-// import { WinnerRewards } from './../services/WinnerRewards';
 import WinnerRewards from '../services/WinnerRewards';
-
-import { useGameContext } from "../services/GameContext";
-
 interface EndScreenProps {
     score: number;
     highestScore?: number;
 }
-
-// const PaymentSection: React.FC = () => {
-//   const {
-//     setHasPaid,
-//     isProcessing,
-//     setIsProcessing,
-//     // paymentError,
-//     setPaymentError,
-//     jwtToken,
-//     inputRef,
-//   } = useGameContext();
-// }
 
 const Container = styled.div`
   padding: 4rem;
@@ -34,7 +17,6 @@ const Container = styled.div`
   background-color: var(--background);
   border-radius: 20px;
   gap: 1rem;
-
 `;
 
 const ScoreItem = styled.div`
@@ -43,7 +25,6 @@ const ScoreItem = styled.div`
   flex-direction: row;
   align-items: center;
   gap:3rem;
-
 `;
 
 const ButtonContainer = styled.div`
@@ -52,20 +33,14 @@ const ButtonContainer = styled.div`
   gap: 1rem;
 `;
 
-interface WinnerRewardsProps {
-  onRewardClaimed: () => void;
-}
-
-
-
 export default function EndScreen({ score }: EndScreenProps) {
   const { highestScore } = useHighScore();
 
     const getResultMessage = () => {
-        if (score >= highestScore && highestScore > 0) return "Nytt rekord!";
-        if (score > 0) return "Bra spelat!";
+        if (score >= highestScore && highestScore > 0) return "Grymt spelat! Nytt rekord!";
+        if (score > 300) return "Grattis du vann! Bra spelat! ";
         console.log(highestScore)
-        return "Försök igen!";
+        return "Bättre lycka nästa gång!";
     };
 
     return (
@@ -76,19 +51,15 @@ export default function EndScreen({ score }: EndScreenProps) {
               <h3>Resultat: {score}</h3>
               <h3>High Score: {highestScore}</h3>
             </ScoreItem>
-         
-          
           
             <h5>{getResultMessage()}</h5>
           
+          <WinnerRewards onRewardClaimed={() => {}} score={score}/>
           <ButtonContainer>
             <Button to="/leaderboard">
               Spara resultat
             </Button>
           </ButtonContainer>
-
-           <WinnerRewards onRewardClaimed={() => {}} score={score}/>
-
   
         </Container>
     );
