+++ src/app/layout.tsx
@@ -1,28 +1,5 @@
-import { Metadata } from 'next';
 import GlobalStyles from './styles/GlobalStyles';
-import { GameProvider } from './services/GameContext';
-
-export const metadata: Metadata = {
-  title: 'RuneHunt',
-  description: 'Rune Hunt Game',
-};
-
-// export const metadata = {
-//   title: 'Next.js',./styles/GlobalStyles
-//   description: 'Generated by Next.js',
-// }
-
-// export default function RootLayout({
-//   children,
-// }: {
-//   children: React.ReactNode
-// }) {
-//   return (
-//     <html lang="en">
-//       <body>{children}</body>
-//     </html>
-//   )
-// }
+import { GameContextProvider } from './services/GameContextProvider';
 
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
@@ -36,11 +13,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
         />
       </head>
       <body>
-        <GameProvider>
-
+        <GameContextProvider>
         <GlobalStyles />
         {children}
-        </GameProvider>
+        </GameContextProvider>
       </body>
     </html>
   );
