+++ src/app/oldPages/RuneHuntPage.tsx
@@ -1,9 +1,6 @@
-
 // import OrientationOverlay from '../components/OrientationCheck/OrientationCheck';
 import Payment from '../components/Payment';
 import JwtDisplay from '../services/JwtDisplay';
-import { GameProvider } from '../services/GameContext';
-
 
 function RuneHuntPage() {
   return (
