+++ src/app/services/JwtDisplay.tsx
@@ -1,14 +1,10 @@
 'use client'
 import { useEffect, useState } from "react";
-import { useGameContext } from "./GameContext";
+import { useGameContext } from "./useGameContext";
 import { decodeJwt } from "./auth";
 
-/**
- * Component to receive and display JWT token from parent application
- */
 export default function JwtDisplay() {
   const { jwtToken, setJwtToken } = useGameContext();
-  const [decodedToken, setDecodedToken] = useState<any>(null);
   const [error, setError] = useState<string | null>(null);
 
   useEffect(() => {
@@ -38,7 +34,6 @@ export default function JwtDisplay() {
           console.log("JWT token set in context:", token);
           const decoded = decodeJwt(token);
           if (decoded) {
-            setDecodedToken(decoded);
             setError(null);
           } else {
             setError("Failed to decode token");
@@ -65,13 +60,7 @@ export default function JwtDisplay() {
   }, [setJwtToken]);
 
   if (!jwtToken) {
-    return (
-      <div className="p-4 bg-yellow-50 border border-yellow-200 rounded-lg mb-6">
-        <p className="text-sm text-yellow-700">
-          Waiting for JWT token from parent application...
-        </p>
-      </div>
-    );
+    return null;
   }
 
   if (error) {
@@ -82,38 +71,5 @@ export default function JwtDisplay() {
     );
   }
 
-  return (
-    <div className="p-4 bg-white rounded-lg border-2 border-[#e0d5c5] mb-6">
-      <h2 className="text-xl font-bold mb-3">JWT Token Information</h2>
-
-      <div className="mb-4">
-        <h3 className="text-sm font-medium text-gray-500 mb-1">Raw Token:</h3>
-        <div className="bg-gray-50 p-3 rounded overflow-auto max-h-24">
-          <p className="text-xs font-mono break-all">{jwtToken}</p>
-        </div>
-      </div>
-
-      {decodedToken && (
-        <div>
-          <h3 className="text-sm font-medium text-gray-500 mb-1">
-            Decoded Content:
-          </h3>
-          <div className="bg-gray-50 p-3 rounded overflow-auto max-h-80">
-            <pre className="text-xs font-mono whitespace-pre-wrap">
-              {JSON.stringify(decodedToken, null, 2)}
-            </pre>
-          </div>
-
-          {decodedToken.exp && (
-            <div className="mt-3 text-xs">
-              <p>
-                <span className="font-medium">Expires:</span>{" "}
-                {new Date(decodedToken.exp * 1000).toLocaleString()}
-              </p>
-            </div>
-          )}
-        </div>
-      )}
-    </div>
-  );
+  return null;
 }
