+++ src/app/services/useGameContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { GameContext } from "./GameContext";
+
+export const useGameContext = () => {
+  const context = useContext(GameContext);
+  if (!context) {
+    throw new Error("useGameContext must be used within a GameProvider");
+  }
+  return context;
+};
