+++ src/app/services/vite.config.ts
@@ -1,4 +1,3 @@
-// vite.config.ts
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
@@ -9,17 +8,17 @@ export default defineConfig({
       '^/api/.*': {
         target: 'https://tivoli.yrgobanken.vip',
         changeOrigin: true,
-        configure: (proxy, _options) => {
-          proxy.on('error', (err, _req, _res) => {
-            console.log('proxy error', err);
-          });
-          proxy.on('proxyReq', (proxyReq, req, _res) => {
-            console.log('Sending Request:', req.method, req.url);
-          });
-          proxy.on('proxyRes', (proxyRes, req, _res) => {
-            console.log('Received Response:', proxyRes.statusCode, req.url);
-          });
-        },
+          configure: (proxy) => {
+            proxy.on('error', (err) => {
+              console.log('proxy error', err);
+            });
+            proxy.on('proxyReq', (proxyReq) => {
+              console.log('Sending Request:', proxyReq.method, proxyReq.path);
+            });
+            proxy.on('proxyRes', (proxyRes) => {
+              console.log('Received Response:', proxyRes.statusCode);
+            });
+          },
       }
     }
   }
